+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.3.1] - 2019-03-09
+
+### Added
+
+* `adjuster.number().map()`
+* `adjuster.string().map()`
+* `adjuster.numericString().map()`
+
 ## [1.3.0] - 2019-01-26
 
 ### Others
@@ -330,7 +338,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * First release.
 
-[Unreleased]: https://github.com/shimataro/node-adjuster/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/shimataro/node-adjuster/compare/v1.3.1...HEAD
+[1.3.1]: https://github.com/shimataro/node-adjuster/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/shimataro/node-adjuster/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/shimataro/node-adjuster/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/shimataro/node-adjuster/compare/v1.0.1...v1.1.0